+++ framework/smd-spring-boot-starter/src/integrationTest/java/app/dodb/smd/spring/metadata/example/AccountAuditRecordedEvent.java
@@ -1,0 +1,12 @@
+package app.dodb.smd.spring.metadata.example;
+
+import app.dodb.smd.api.event.Event;
+
+import static java.util.Objects.requireNonNull;
+
+public record AccountAuditRecordedEvent(String name) implements Event {
+
+    public AccountAuditRecordedEvent {
+        requireNonNull(name);
+    }
+}