+++ framework/smd-spring-boot-starter/src/integrationTest/java/app/dodb/smd/spring/metadata/example/LookupAccountAuditQuery.java
@@ -1,0 +1,6 @@
+package app.dodb.smd.spring.metadata.example;
+
+import app.dodb.smd.api.query.Query;
+
+public record LookupAccountAuditQuery() implements Query<Integer> {
+}