+++ framework/smd-spring-boot-starter/src/integrationTest/java/app/dodb/smd/spring/metadata/example/MetadataExampleIntegrationTestConfiguration.java
@@ -1,0 +1,9 @@
+package app.dodb.smd.spring.metadata.example;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan
+public class MetadataExampleIntegrationTestConfiguration {
+}